Complete in  Essential Oil and demo to Client

+++ app/Http/Controllers/EssentialOilTypeController.php
@@ -4,17 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Models\EssentialOilType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class EssentialOilTypeController extends Controller
 {
     public function store(Request $request)
     {
+        $fileImage = $request->file('fileImage');
         $name = $request->get('name');
+        $idType = time();
 
         $essentialOilType = new EssentialOilType();
-        $essentialOilType->id = time();
+        $essentialOilType->id = $idType;
         $essentialOilType->EssentialOilType_Name = $name;
+        $essentialOilType->EssentialOilType_Image = $idType;
         $essentialOilType->save();
+
+        Storage::disk('local')->putFileAs(
+            'public/images/essential-oil/type/' . $idType,
+            $fileImage,
+            $idType . '.png');
 
         echo json_encode(['status' => 200, 'message' => 'ok']);
     }
@@ -38,12 +48,49 @@ class EssentialOilTypeController extends Controller
 
     public function delete(Request $request){
         $id = $request->get('id');
-        $typeDelete = EssentialOilType::where('id', $id)->get()->first();
-        $typeDelete->delete();
+        $response = [
+            'status' => 200,
+            'message' => ''
+        ];
+
+        $categoryHasThisID = DB::table('essential_oil_categories')
+            ->where(
+                'FkEssentialOilType_id',
+                '=',
+                $id
+            )
+            ->get()->count();
+
+        if($categoryHasThisID == 0){
+            $typeDelete = EssentialOilType::where('id', $id)->first();
+            $typeDelete->delete();
+        }else{
+            $response['status'] = 303;
+        }
+
+        echo json_encode($response);
     }
 
     public function getAll()
     {
-        echo json_encode(['status' => 200, 'data' => EssentialOilType::all()]);
+        $dataMain = DB::table('essential_oil_types')
+            ->select()->get();
+
+        foreach ($dataMain as $item) {
+            $count = DB::table('essential_oil_categories')
+                ->select([
+                    DB::raw('COUNT(essential_oil_categories.FkEssentialOilType_id) as category')
+                ])
+                ->where(
+                    'FkEssentialOilType_id',
+                    '=',
+                    $item->id
+                )
+                ->get();
+
+            $item->count = $count;
+        }
+
+        echo json_encode(['status' => 200, 'data' => $dataMain]);
     }
 }
